[image_picker] Mention `launchMode: singleInstance` in README (#3759)
Users of the image picker are reporting issues for Android when setting the `launchMode` of the application activity to `singleInstance`. Namely, the image picker will always return `RESULT_CANCELED`. After investigation, it turned out that this problem is inherent by the way Android handles activities with this launch mode: `singleInstance` instructs the framework to open new activities in a new task. As activities cannot communicate between tasks, the image picker activity is unable to relay back the result to the calling activity.

This PR updates the README to mention this behavior and provide a workaround.

Closes [76856](https://github.com/flutter/flutter/issues/76856).